Adds the ability to chunk the listing of content items within a space as well as the ability to narrow the list based on a prefix. Note that this is implemented in S3 and RackSpace, but not in EMC as EMC has no facility for requesting a partial listing of content items within a container.
git-svn-id: https://svn.duraspace.org/duracloud-priv/trunk@415 3c0c4179-e91d-42e7-a082-8e2792ce3676

+++ duradmin/src/main/java/org/duracloud/duradmin/util/Scrollable.java
@@ -16,7 +16,7 @@ public interface Scrollable<E> {
      * 
      * @return
      */
-    public int getMaxResultsPerPage();
+    public long getMaxResultsPerPage();
 
     public void setMaxResultsPerPage(int maxResults);
